% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{hex_transparacy}
\alias{hex_transparacy}
\title{Adjust transparency of a hex string}
\usage{
hex_transparacy(hex, percent = 1)
}
\arguments{
\item{hex}{A 6-character hex string (e.g. #000000)}

\item{percent}{Transparency level from 0-1}
}
\value{
A hex string
}
\description{
Adjust transparency of a hex string
}
