% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{riso_color}
\alias{riso_color}
\title{Get a risk ink color}
\usage{
riso_color(color, alpha = 1)
}
\arguments{
\item{color}{Color name}

\item{alpha}{Color transparency}
}
\value{
A hex string
}
\description{
Get a risk ink color
}
