% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{riso_names}
\alias{riso_names}
\title{Riso color names}
\usage{
riso_names(sorted = FALSE)
}
\arguments{
\item{sorted}{Optionally sort values}
}
\value{
A character vector of names
}
\description{
Riso color names
}
