% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{show_colors}
\alias{show_colors}
\title{Plot one or more colors}
\usage{
show_colors(colors, title = "")
}
\arguments{
\item{colors}{Character vector of hex strings}

\item{title}{Optional plot title}
}
\value{
A `gg` object
}
\description{
Plot one or more colors
}
